Fixed Opened Issues

+++ back/healthapp/src/main/java/com/healthcare/domain/service/MedicServiceImpl.java
@@ -1,23 +1,29 @@
 package com.healthcare.domain.service;
 
 import com.healthcare.domain.dto.MedicDTO;
-import com.healthcare.domain.exceptions.NotFoundInDatabaseException;
+import com.healthcare.domain.exceptions.*;
 import com.healthcare.domain.model.entity.Medic;
+import com.healthcare.domain.repository.AppointmentRepository;
 import com.healthcare.domain.repository.MedicRepository;
-import jakarta.transaction.Transactional;
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
+
+import org.modelmapper.ModelMapper;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Map;
 
+@RequiredArgsConstructor
 @Service
 public class MedicServiceImpl implements IMedicService {
 
-    @Autowired
-    private MedicRepository medicRepository;
+
+    private final MedicRepository medicRepository;
+    private final AppointmentRepository appointmentRepository;
+    private final ModelMapper modelMapper;
 
     @Override
     public ResponseEntity<?> getAllMedics(String speciality, String gender, String state) {
@@ -40,7 +46,7 @@ public class MedicServiceImpl implements IMedicService {
         }
 
         List<MedicDTO> medicDTOS = medics.stream()
-                .map(MedicDTO::fromEntity)
+                .map(medicEntity -> modelMapper.map(medicEntity, MedicDTO.class))
                 .toList();
 
         if (medicDTOS.isEmpty()) {
@@ -53,25 +59,40 @@ public class MedicServiceImpl implements IMedicService {
     @Override
     public ResponseEntity<?> getMedicById(Long id) {
         Medic medic = medicRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Médico no encontrado con ID: " + id));
-        MedicDTO medicDTO = MedicDTO.fromEntity(medic);
-        return ResponseEntity.ok(Map.of("medic", medicDTO));
+                .orElseThrow(() -> new MedicNotFoundException("Médico no encontrado"));
+        MedicDTO dto = modelMapper.map(medic, MedicDTO.class);
+        return ResponseEntity.ok(Map.of("medic", dto));
     }
 
     @Override
     @Transactional
-    public ResponseEntity<?> createMedic(Medic medic) {
-        Medic savedMedic = medicRepository.save(medic);
-        return new ResponseEntity<>(Map.of(
-                "message", "Médico creado con éxito", "medic", savedMedic),
-                HttpStatus.CREATED);
+    public ResponseEntity<?> createMedic(MedicDTO medicDTO) {
+        if(medicDTO == null){
+            throw new InvalidDataException("El cuerpo de la solicitud es obligatorio");
+        }
+
+        if(medicRepository.existsByDocumentId(medicDTO.getDocumentId())) {
+            throw new DuplicatedEntryEx("Médico ya registrado");
+        }
+
+        Medic medic = new Medic(medicDTO);
+        medicRepository.save(medic);
+        return ResponseEntity.status(HttpStatus.CREATED).body(Map.of(
+                "message", "Médico creado con éxito",
+                "medic", medicDTO
+        ));
     }
 
     @Override
     @Transactional
     public ResponseEntity<?> deleteMedic(Long id) {
         Medic medic = medicRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Médico no encontrado con ID: " + id));
+                .orElseThrow(() -> new MedicNotFoundException("Médico no encontrado"));
+
+        if(appointmentRepository.existsByMedicId(id)) {
+            throw new MedicDeletionException("No se puede eliminar el Médico porque tiene citas asociadas");
+        }
+
         medicRepository.delete(medic);
         return ResponseEntity.ok(Map.of("message", "Médico eliminado con éxito"));
     }
